Add MVP for a pipe busting trampoline for commands wrapcmd.
This is a more general form of being able to wrap a command to allow for
hook functions to determine if something is/n't good.

So after importing sh/lib.sh one can simply issue:

wrapcmd command-a command-b command-c

And if DEBUG is set in the environment, you will get the wrapped
behavior instead of calling commands directly.

This setup is FULL of caveats, aka don't wrap stuff that wants input,
with this set everything will be slower (it literally breaks a pipe),
and there is a ton of other caveats like the output will be different
than if you ran it directly (I separate out stdout and stderr).

But for a POC/MVP its ok. Example wrapper script using this is in the
examples directory.

More to come later.