Update wrapcmd and add a few more tests
Since this is a wrapper, nuking the validate hook. Its point was to enable doing
stuff like jq $file prior to running jq to determine if something is or isn't
json. But with the pipe busting that becomes rather obvious so I don't see the
need or usefulness of this specific hook.

Adds a default post hook that dumps out useful information. The example is
updated to use it.

Added a basic test suite for the internal function setup and also when importing
lib.sh just have it automatically call the cleanup trap for the library saving
callers needing to care/know about the temp dir in use.