MTL-2076 Switch to virtualenv
This switches to using a virtualenv instead of a system installation and
adds GitHub promotion workflows.

The SUSE library support for Python39 and Python310 was very lacking.
Now with a virtualenv we can use any Python library and that library
will get bundled into our virtualenv.

The new virtualenv does not include `pip`, `setuptools`, or `wheel` in
order to shrink its size. These tools are not necessary, if someone
needs to use `pip` against the virtualenv on an installed system
(voiding the RPM install) then they can use `pip install --root=`.