Ensure the deprecated package is uninstalled (#20)
* Ensure the deprecated package is uninstalled

If a system has `pythonXY-libcsm` installed it must be uninstalled. Even
though `pythonXY-libcsm` and `libcsm` both `Provides: libcsm` it is
possible to still install both packages. This package is so new that
that edge case is rare, but it should be addressed.

* Fix virtualenv

The `pip --root` option was not correctly installing the `libcsm` module
into the virtualenv. During testing a false positive was generated
because the module was getting invoked while the tester was running in
the cloned repository.