Fix RPM build for deps using `setuptools`
When I added kubernetes in a test branch the RPM build was
failing, this was because kubernetes was bringing setuptools
in which includes some odd-named files that makes `INSTALLED_FILES`
unparseable. This was circumventing the `--no-setuptools`
option passed to our virtualenv.

This ensures that we always purge build tools, and do not rely on the
`virtualenv` setup to omit them.